Moved OSX demos around
